requirements.txt
discord.py
yt_dlp
ffmpeg-python
asyncio
PyNaCl